# Optimize byte swap base62 encoding