This PR revisits an optimization that was added when we originally released this package, but appears not to be needed anymore:
```
name                    old time/op  new time/op  delta
AppendFastEncodeBase62   174ns ± 3%   165ns ± 3%  -4.91%  (p=0.000 n=10+10)
```